Seção 9: DTOs e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,8 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ProEventos.API.Dtos;
 using ProEventos.Application.Contratos;
 using ProEventos.Domain;
 
@@ -25,7 +27,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado.");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -43,7 +45,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null) return NotFound("Evento por id não encontrado.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -61,7 +63,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(evento == null) return NotFound("Eventos por tema não encontrados.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -74,13 +76,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {           
             try
             {
                 var eventos = await _eventoService.AddEventos(model);
-                if(eventos == null) return BadRequest("Erro ao tentar adicionar evento.");
-
+                if(eventos == null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -92,12 +93,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {           
             try
             {
                 var eventos = await _eventoService.UpdateEvento(id, model);
-                if(eventos == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -114,9 +115,11 @@ namespace ProEventos.API.Controllers
         {           
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento == null) return NoContent();
                 return await _eventoService.DeleteEvento(id) ? 
                         Ok("Deletado") :
-                        BadRequest("Evento não deletado");
+                        throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento. ");
             }
             catch (Exception ex)
             {
